pyproject.toml
[tool.poetry]
name = "phishing-service"
version = "0.1.0"
description = ""
authors = ["Emre Eren <emreren@hotmail.com>"]
readme = "README.md"
packages = [{include = "phishing_service"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.0.2"
httpx = "^0.24.1"
sqlalchemy = "^2.0.18"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"
fastapi = "^0.100.0"
uvicorn = "^0.23.0"
validators = "^0.20.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
